feat: make it so `testFunction` can throw errors (#86)
* testFunction can throw

Previously, running the MockService only took a non-throwing
closure. This is place in which testers will place their networking
requests. It is common for networking requests to throw, however. Not
being able to throw in the test function meant testers had to catch and
fail their tests on their own:

    myMockService.run(timeout: 10) { (testComplete) -> Void in
      do {
        try serviceClientUnderTest.getUsers( /* ... */ )
      } catch {
        XCTFail("error thrown: \(error.localizedDescription)")
      }
    }

It would be cleaner if `run` would accept a throwing function and fail
the test itself. This would become:

    myMockService.run(timeout: 10) { (testComplete) -> Void in
      try serviceClientUnderTest.getUsers( /* ... */ )
    }

[NB: The `run` function does not rethrow the error.]

* A few PR review changes

PR: DiUS/pact-consumer-swift/pull/86