build: Truncate commit message when triggering example project builds to avoid curl error (#111)
GitHub Actions env `github.event.head_commit.message` takes the complete commit message. When passing it to `curl` command, the curl command fails when message is multiline.

With this only the first line of the commit message is used and passed to the demo projects’ GitHub Actions build run name (`event_type`).